+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 - [fix: remove conflict dependencies management](https://github.com/Tencent/spring-cloud-tencent/pull/1683)
 - [refactor: optimize log](https://github.com/Tencent/spring-cloud-tencent/pull/1684)
 - [feat:support setting load balancing strategy per service.](https://github.com/Tencent/spring-cloud-tencent/pull/1685)
+- [feat: support shortest response time lb and least connection lb](https://github.com/Tencent/spring-cloud-tencent/pull/1686)